LearningBot/requirements.txt
# Основные зависимости для Telegram бота
aiogram==3.13.1
aiohttp==3.10.10

# База данных и ORM
SQLAlchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0  # PostgreSQL драйвер
aiosqlite==0.20.0  # SQLite драйвер для разработки

# Конфигурация и переменные окружения  
python-decouple==3.8
pydantic>=2.4.1,<2.10
pydantic-settings==2.6.1

# Кэширование (опционально)
redis==5.2.0
aioredis==2.0.1

# Утилиты для работы с датами и временем
python-dateutil==2.9.0

# Логирование и мониторинг
structlog==24.4.0

# Тестирование
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Линтеры и форматтеры (для разработки)
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Работа с изображениями (для обработки медиа)
Pillow==11.0.0

# Дополнительные утилиты
httpx==0.27.2